+++ tests/algorithms/pytorch/test_newton_raphson.py
@@ -10,12 +10,18 @@ from connectlib.dependency import Dependency
 from connectlib.evaluation_strategy import EvaluationStrategy
 from connectlib.exceptions import CriterionReductionError
 from connectlib.exceptions import NegativeHessianMatrixError
+from connectlib.model_loading import download_algo_files
+from connectlib.model_loading import load_algo
 from connectlib.nodes.test_data_node import TestDataNode
 from connectlib.nodes.train_data_node import TrainDataNode
 from connectlib.strategies import NewtonRaphson
 from tests import utils
 
 from ... import assets_factory
+
+# For a convex problem and with damping_factor = 1, Newton Raphson is supposed to reach the global optimum
+# in one round.
+EXPECTED_PERFORMANCE = 0
 
 
 @pytest.fixture(scope="module")
@@ -208,17 +214,29 @@ def test_train_newton_raphson_non_convex_cnn(torch_algo):
         my_algo.train(x=x_train, y=y_train, _skip=True)
 
 
-@pytest.mark.substra
-@pytest.mark.slow
-def test_pytorch_newton_raphson_algo_performance(
+@pytest.fixture(scope="module")
+def nr_test_data():
+    return [np.array([[5, 5, 11], [6, 6, 13], [7, 7, 15]])]
+
+
+@pytest.fixture(scope="module")
+def compute_plan(
     network,
+    nr_test_data,
     numpy_datasets,
     perceptron,
     aggregation_node,
+    mae,
     session_dir,
-    default_permissions,
 ):
-    """End to end test for torch Newton Raphson algorithm."""
+    """Compute plan for e2e test"""
+
+    damping_factor = 1
+    num_rounds = 1
+    batch_size = 1
+
+    seed = 42
+    torch.manual_seed(seed)
 
     # We define several sample without noises to be sure to reach the global optimum.
     # Here, f(x1,x2) = x1 + x2 + 1
@@ -238,16 +256,8 @@ def test_pytorch_newton_raphson_algo_performance(
         for k in range(network.n_organizations)
     ]
 
-    metric = assets_factory.add_python_metric(
-        client=network.clients[0],
-        tmp_folder=session_dir,
-        python_formula="((y_pred - y_true)**2).mean()",
-        name="MSE",
-        permissions=default_permissions,
-    )
-
     test_sample_nodes = assets_factory.add_numpy_samples(
-        contents=[np.array([[5, 5, 11], [6, 6, 13], [7, 7, 15]])],
+        contents=nr_test_data,
         dataset_keys=[numpy_datasets[0]],
         clients=[network.clients[0]],
         tmp_folder=session_dir,
@@ -258,20 +268,9 @@ def test_pytorch_newton_raphson_algo_performance(
             network.msp_ids[0],
             numpy_datasets[0],
             [test_sample_nodes[0]],
-            metric_keys=[metric],
+            metric_keys=[mae.key],
         )
     ]
-
-    # For a convex problem and with damping_factor = 1, Newton Raphson is supposed to reach the global optimum
-    # in one round.
-    expected_performance = 0
-
-    damping_factor = 1
-    num_rounds = 1
-    batch_size = 1
-
-    seed = 42
-    torch.manual_seed(seed)
 
     model = perceptron(linear_n_col=2, linear_n_target=1)
     criterion = torch.nn.MSELoss()
@@ -317,7 +316,38 @@ def test_pytorch_newton_raphson_algo_performance(
     # Wait for the compute plan to be finished
     utils.wait(network.clients[0], compute_plan)
 
+    return compute_plan
+
+
+@pytest.mark.substra
+@pytest.mark.slow
+def test_pytorch_nr_algo_performance(
+    network,
+    compute_plan,
+):
+
     perfs = network.clients[0].get_performances(compute_plan.key)
 
-    rel = 1e-6  # This relative error is due to the l2 regularization, mandatory to reach numerical stability.
-    assert pytest.approx(expected_performance, abs=rel) == perfs.performance[0]
+    rel = 1e-5
+    # This abs_ative error is due to the l2 regularization, mandatory to reach numerical stability.
+    # This fails on mac M1 pro with 1e-5
+    # TODO investigate
+    assert pytest.approx(EXPECTED_PERFORMANCE, abs=rel) == perfs.performance[0]
+
+
+@pytest.mark.slow
+@pytest.mark.substra
+def test_download_load_algo(network, compute_plan, session_dir, nr_test_data, mae):
+    download_algo_files(
+        client=network.clients[0], compute_plan_key=compute_plan.key, round_idx=None, dest_folder=session_dir
+    )
+    model = load_algo(input_folder=session_dir)._model
+
+    y_pred = model(torch.from_numpy(nr_test_data[0][:, :-1]).float()).detach().numpy().reshape(-1)
+    y_true = nr_test_data[0][:, -1:].reshape(-1)
+    performance = mae.compute(y_pred, y_true)
+
+    # This test fails with default approx parameters
+    # TODO: investigate
+    rel = 1e-5
+    assert performance == pytest.approx(EXPECTED_PERFORMANCE, abs=rel)