+++ tests/algorithms/pytorch/test_scaffold.py
@@ -14,6 +14,8 @@ from connectlib.evaluation_strategy import EvaluationStrategy
 from connectlib.exceptions import NumUpdatesValueError
 from connectlib.exceptions import TorchScaffoldAlgoParametersUpdateError
 from connectlib.index_generator import NpIndexGenerator
+from connectlib.model_loading import download_algo_files
+from connectlib.model_loading import load_algo
 from connectlib.schemas import ScaffoldAveragedStates
 from connectlib.schemas import ScaffoldSharedState
 from connectlib.strategies import Scaffold
@@ -24,6 +26,8 @@ from tests.algorithms.pytorch.torch_tests_utils import assert_tensor_list_not_ze
 
 logger = logging.getLogger(__name__)
 current_folder = Path(__file__).parent
+
+EXPECTED_PERFORMANCE = 0.0127768706
 
 
 def _torch_algo(torch_linear_model, lr=0.1, use_scheduler=False):
@@ -159,11 +163,9 @@ def test_pytorch_scaffold_algo_performance(
 ):
     """End to end test for torch fed avg algorithm."""
 
-    expected_performance = 0.0127768706
-
     testtuples = network.clients[0].list_testtuple(filters={"compute_plan_key": [compute_plan.key]})
     testtuple = testtuples[0]
-    assert list(testtuple.test.perfs.values())[0] == pytest.approx(expected_performance, rel=rtol)
+    assert list(testtuple.test.perfs.values())[0] == pytest.approx(EXPECTED_PERFORMANCE, rel=rtol)
 
 
 @pytest.mark.skip("Can't use inference mode because of the set weight function")
@@ -429,3 +431,18 @@ def test_update_parameters_call(nb_update_params_call, torch_linear_model, num_u
         my_algo.train(x=np.random.random(10), y=np.random.random(10), _skip=True)
 
     assert my_algo._scaffold_parameters_update_num_call == nb_update_params_call
+
+
+@pytest.mark.slow
+@pytest.mark.substra
+def test_download_load_algo(network, compute_plan, session_dir, test_linear_data_samples, mae):
+    download_algo_files(
+        client=network.clients[0], compute_plan_key=compute_plan.key, round_idx=None, dest_folder=session_dir
+    )
+    model = load_algo(input_folder=session_dir)._model
+
+    y_pred = model(torch.from_numpy(test_linear_data_samples[0][:, :-1]).float()).detach().numpy().reshape(-1)
+    y_true = test_linear_data_samples[0][:, -1:].reshape(-1)
+    performance = mae.compute(y_pred, y_true)
+
+    assert performance == pytest.approx(EXPECTED_PERFORMANCE)